requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --emit-index-url
--index-url https://pypi.org/simple
--extra-index-url https://us-central1-python.pkg.dev/cds-artifacts/public-python/simple/

annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
ariadne-codegen==0.14.0
    # via depmap-omics-wgs (pyproject.toml)
attrs==25.3.0
    # via hypothesis
autoflake==2.3.1
    # via ariadne-codegen
black==25.1.0
    # via ariadne-codegen
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   ariadne-codegen
    #   black
    #   flask
    #   functions-framework
    #   typer
cloudevents==1.12.0
    # via functions-framework
cryptography==45.0.5
    # via pyjwt
deprecated==1.2.18
    # via pygithub
deprecation==2.1.0
    # via cloudevents
firecloud-api-cds==0.2.0
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   nebelung
flask==3.1.1
    # via functions-framework
functions-framework==3.8.3
    # via depmap-omics-wgs (pyproject.toml)
google-api-core==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   firecloud-api-cds
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.12.1
    # via depmap-omics-wgs (pyproject.toml)
google-cloud-secret-manager==2.24.0
    # via depmap-omics-wgs (pyproject.toml)
google-cloud-storage==3.2.0
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   firecloud-api-cds
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
graphql-core==3.2.6
    # via ariadne-codegen
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.73.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.73.1
    # via google-api-core
gunicorn==23.0.0
    # via functions-framework
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ariadne-codegen
hypothesis==6.135.32
    # via pandera
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
isort==6.0.1
    # via ariadne-codegen
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
multimethod==1.10
    # via pandera
mypy-extensions==1.1.0
    # via
    #   black
    #   typing-inspect
nebelung==2.9.4
    # via depmap-omics-wgs (pyproject.toml)
numpy==2.3.1
    # via
    #   pandas
    #   pandera
opentelemetry-api==1.35.0
    # via google-cloud-logging
packaging==25.0
    # via
    #   black
    #   deprecation
    #   gunicorn
    #   pandera
pandas==2.3.1
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   nebelung
    #   pandera
    #   pd-flatten
pandera==0.20.4
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   nebelung
pathspec==0.12.1
    # via black
pd-flatten==0.1.2
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   nebelung
platformdirs==4.3.8
    # via black
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   ariadne-codegen
    #   nebelung
    #   pandera
pydantic-core==2.33.2
    # via pydantic
pydot==4.0.1
    # via firecloud-api-cds
pyflakes==3.4.0
    # via autoflake
pygithub==2.6.1
    # via
    #   depmap-omics-wgs (pyproject.toml)
    #   nebelung
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.2.3
    # via pydot
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via depmap-omics-wgs (pyproject.toml)
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-storage
    #   pygithub
rich==14.0.0
    # via typer
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   firecloud-api-cds
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sortedcontainers==2.4.0
    # via hypothesis
toml==0.10.2
    # via ariadne-codegen
tqdm==4.67.1
    # via depmap-omics-wgs (pyproject.toml)
typeguard==4.4.4
    # via pandera
typer==0.16.0
    # via depmap-omics-wgs (pyproject.toml)
typing-extensions==4.14.1
    # via
    #   anyio
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   typeguard
    #   typer
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via pandera
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   pygithub
    #   requests
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.3
    # via
    #   flask
    #   functions-framework
wrapt==1.17.2
    # via
    #   deprecated
    #   pandera
zipp==3.23.0
    # via importlib-metadata
